Adding support for repl
